starting to work on event as document option for mongodb event store
