+++ src/packages/emmett-mongodb/src/eventStore/eventAsDocument/index.ts
@@ -1,0 +1,1 @@
+export * from './mongoDBEventAsDocumentEventStore';
