--- src/packages/emmett-mongodb/src/eventStore/index.ts
@@ -1,3 +1,0 @@
-export * from './mongoDBEventStore';
-export * from './projections';
-export * from './storage';
