+++ src/packages/emmett-mongodb/src/eventStore/streamAsDocument/index.ts
@@ -1,0 +1,3 @@
+export * from './mongoDBStreamAsDocumentEventStore';
+export * from './projections';
+export * from './storage';
