+++ src/packages/emmett-mongodb/src/eventStore/streamAsDocument/mongoDBStreamAsDocumentEventStore.connection.e2e.spec.ts
