+++ src/packages/emmett-mongodb/src/eventStore/streamAsDocument/mongoDBStreamAsDocumentEventStore.storage.e2e.spec.ts
