+++ src/packages/emmett-mongodb/src/eventStore/streamAsDocument/mongoDBStreamAsDocumentEventStore.unit.spec.ts
