+++ src/packages/emmett-mongodb/src/eventStore/streamAsDocument/projections/index.ts
