+++ src/packages/emmett-mongodb/src/eventStore/streamAsDocument/projections/mongoDBEventStore.projections.e2e.spec.ts
@@ -19,7 +19,7 @@ import {
   toStreamCollectionName,
   toStreamName,
   type EventStream,
-  type MongoDBEventStore,
+  type MongoDBStreamAsDocumentEventStore,
 } from '../';
 import {
   type DiscountApplied,
@@ -30,7 +30,7 @@ import {
 
 void describe('MongoDBEventStore', () => {
   let mongodb: StartedMongoDBContainer;
-  let eventStore: MongoDBEventStore;
+  let eventStore: MongoDBStreamAsDocumentEventStore;
   let client: MongoClient;
   let collection: Collection<EventStream>;
