+++ src/packages/emmett-mongodb/src/eventStore/streamAsDocument/projections/mongoDBInlineProjection.ts
