+++ src/packages/emmett-mongodb/src/eventStore/streamAsDocument/projections/mongoDBInlineProjectionSpec.single.int.spec.ts
