+++ src/packages/emmett-mongodb/src/eventStore/streamAsDocument/projections/mongoDBInlineProjectionSpec.ts
