+++ src/packages/emmett-mongodb/src/eventStore/streamAsDocument/storage/index.ts
