+++ src/packages/emmett-mongodb/src/eventStore/streamAsDocument/storage/mongoDBEventStoreStorage.unit.spec.ts
